Towards modifier inversion: pass target up
The top CO of the CO ("Head", slight misnomer, should change that) is
passed down as a default. On every instance of modification, the
argument top CO is passed down instead. The top CO that makes it to the
bottom is passed back up.

TODO: implement maplist/6 and use the top COs to fix arguments for other
tokens.